Promote the active filters component out of Find
